1/service_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.11
pydantic==2.8.2
boto3==1.34.69
pika==1.3.2
python-multipart==0.0.9
requests==2.32.3
python-dotenv